+++ src/MwcAppDialog.ts
@@ -17,8 +17,7 @@ export class MwcAppDialog extends LitElement {
   @property({ type: String })
   cancelText = 'cancel'
 
-  @property({ attribute: false })
-  onClose?: Function
+  protected onClose: Function = () => {}
 
   @query('#dialog')
   dialog!: Dialog
@@ -31,18 +30,18 @@ export class MwcAppDialog extends LitElement {
 
   render() {
     return html`
-    <mwc-dialog id="dialog" title="${this.title}" @closing  ="${this.onClose}">
+    <mwc-dialog id="dialog" title="${this.title}" @closed="${(e:CustomEvent) => this.onClose(e)}">
       <div id="content">${this.template}</div>
       ${this._cancelButton ? html`
       <mwc-button slot="secondaryAction" dialogAction="cancel">${this.cancelText}</mwc-button>
       ` : null}
       ${this._acceptButton ? html`
-      <mwc-button unelevated slot="primaryAction" dialogAction="accept">${this.acceptText}</mwc-button>` : null}
+      <mwc-button unelevated slot="primaryAction" @click="${() => this.onClose()}">${this.acceptText}</mwc-button>` : null}
     </mwc-dialog>
     `
   }
 
-  async _open(title: string, template: TemplateResult|string, onAccept: Function, onCancel: Function, cancelButton: boolean = true, acceptButton: boolean = true) {
+  async _open(title: string, template: TemplateResult|string, onAccept: Function, onCancel: Function, acceptButton: boolean = true, cancelButton: boolean = true) {
     
     this.title = title && title[0].toUpperCase() + title.slice(1, title.length)
     this.template = template
@@ -56,18 +55,28 @@ export class MwcAppDialog extends LitElement {
       this.content[el.id] = el
     })
 
-    this.onClose = (e:CustomEvent) => {
-      if ((e.detail.action === 'cancel' || e.detail.action === 'close') && onCancel) {
-        onCancel()
-      } else if (e.detail.action === 'accept' && onAccept) {
-        onAccept(this.content)
-      } else if (onAccept) {
-        onAccept(this.content, e.detail.action)
+    this.onClose = async (e:CustomEvent|undefined) => {
+      if (e) {
+        if ((e.detail.action === 'cancel' || e.detail.action === 'close') && onCancel) {
+          onCancel()
+        } else if (onAccept) {
+          onAccept(this.content, e.detail.action)
+        }
+      }
+      else {
+        // the button was pressed
+        if (onAccept) {
+          const returnCode = await onAccept(this.content)
+          if (returnCode !== -1) {
+            this.onClose = () => {}
+            this.dialog.open = false
+          }
+        }
       }
 
       /* reinit the buttons */
-      this._acceptButton = true
-      this._cancelButton = true
+      // this._acceptButton = true
+      // this._cancelButton = true
     }
     return this.content
   }
@@ -76,40 +85,35 @@ export class MwcAppDialog extends LitElement {
     this.dialog.open = false
   }
 
-  open(title: string, template: TemplateResult|string, onAccept: Function, onCancel: Function|boolean = true) {
+  async open(title: string, template: TemplateResult|string, onAccept:Function|undefined, onCancel: Function|boolean = true) {
     if (typeof template === 'string') {
       template = html`${template}`
     }
-
-    return new Promise((resolve, reject) => {
-      this._open(
+    
+    return await this._open(
         title,
         <TemplateResult>template,
         async (dom:HTMLElement, value:any) => {
           if (onAccept) {
-            const result = await onAccept(dom, value)
-            if (result === -1) {
-              this.dialog.open=true
-              // return
-            }
+            return await onAccept(dom, value)
           }
-          resolve({ dom, value })
+          // _resolve({ dom, value })
         },
         () => {
           if (typeof onCancel === 'function') {
             onCancel()
           }
-          reject()
+          // _reject()
         },
-        typeof onCancel === 'boolean' ? onCancel : onCancel !== undefined,
-        onAccept !== undefined
+        onAccept !== undefined,
+        typeof onCancel === 'boolean' ? onCancel : onCancel !== undefined
       )
-    })
+
   }
 
   notice(title: string, template: TemplateResult | string) {
-    return new Promise((resolve, reject) => {
-      this._open(title, template, (dom: HTMLElement) => resolve(dom), reject, false, true)
+    return new Promise((resolve) => {
+  this._open(title, template, resolve, resolve, true, false)
     })
   }
 
@@ -131,7 +135,7 @@ export class MwcAppDialog extends LitElement {
     `
 
     return new Promise<any>((resolve, reject) => {
-      this._open(title, template, (_:HTMLElement, value:any) => resolve(value), reject, true, false)
+      this._open(title, template, (_:HTMLElement, value:any) => resolve(value), reject, false, true)
     })
   }
 }